fix bug

+++ ali_dindin/models/partner.py
@@ -120,7 +120,7 @@ class ResPartner(models.Model):
         for res in self:
             din_userid = res.din_userid
             super(ResPartner, self).unlink()
-            if self.env['ir.config_parameter'].sudo().get_param('ali_dindin.din_delete_extcontact'):
+            if self.env['ir.config_parameter'].sudo().get_param('ali_dindin.din_delete_extcontact') and din_userid:
                 self.delete_din_extcontact(din_userid)
             return True
 
@@ -134,6 +134,79 @@ class ResPartner(models.Model):
         except Exception as e:
             raise UserError(e)
 
+
+    @api.multi
+    def get_dingding_partner(self):
+        """
+        外部联系人详情
+        获取企业外部联系人详情
+        :param user_id: userId
+        返回数据结构：
+        {'ding_open_errcode': 0, 
+        'result': {
+            'address': '新世纪大厦', 
+            'company_name': '新时代电脑', 
+            'follower_user_id': '01454209426971', 
+            'label_ids': {'number': [136234018, 136234017, 136234029]}, 
+            'mobile': '18968897668', 
+            'name': '曾为龙', 
+            'remark': '电脑维保', 
+            'share_dept_ids': {'number': [108280604]}, 
+            'share_user_ids': {}, 
+            'state_code': '86', 
+            'title': '总经理', 
+            'userid': '014711380725999389'}, 
+        'success': True}
+
+        """
+        for partner in self:
+            userid = partner.din_userid
+            try:
+                client = get_client(self)
+                result = client.tbdingding.dingtalk_corp_extcontact_get(userid)
+                logging.info(">>>获取外部联系人返回结果:{}".format(result))
+                if result.get('ding_open_errcode') == 0:
+                    result = result.get('result')
+                    # 获取标签
+                    label_list = list()
+                    for label in result.get('label_ids').get('number'):
+                        category = self.env['res.partner.category'].sudo().search(
+                            [('din_id', '=', label)])
+                        if category:
+                            label_list.append(category[0].id)
+                    data = {
+                        'name': result.get('name'),
+                        'function': result.get('title'),
+                        'category_id': [(6, 0, label_list)],  # 标签
+                        'din_userid': result.get('userId'),  # 钉钉用户id
+                        'comment': result.get('remark'),  # 备注
+                        'street': result.get('address'),  # 地址
+                        'mobile': result.get('mobile'),  # 手机
+                        'phone': result.get('mobile'),  # 电话
+                        'din_company_name': result.get('company_name'),  # 钉钉公司名称
+                    }
+                    # 获取负责人
+                    if result.get('follower_user_id'):
+                        follower_user = self.env['hr.employee'].sudo().search(
+                            [('din_id', '=', result.get('follower_user_id'))])
+                        data.update({'din_employee_id': follower_user[0].id if follower_user else ''})
+                    # 获取共享范围
+                    if result.get('shareDeptIds'):
+                            dep_din_ids = result.get('shareDeptIds')
+                            dep_list = self.env['hr.department'].sudo().search([('din_id', 'in', dep_din_ids)])
+                            data.update({'din_share_department_ids': [(6, 0, dep_list.ids)] if dep_list else ''})
+                    # 获取共享员工
+                    if result.get('shareUserIds'):
+                            emp_din_ids = result.get('shareUserIds')
+                            emp_list = self.env['hr.employee'].sudo().search([('din_id', 'in', emp_din_ids)])
+                            data.update({'din_share_employee_ids': [(6, 0, emp_list.ids)] if emp_list else ''})
+                    # 根据userid查询联系人是否存在
+                    partner.sudo().write(data)
+                else:
+                    _logger.info("从钉钉同步联系人时发生意外，原因为:{}".format(result.get('errmsg')))
+                    partner.message_post(body="从钉钉同步联系人失败:{}".format(result.get('errmsg')), message_type='notification')
+            except Exception as e:
+                raise UserError(e)
 
 # 未使用，但是不能删除，因为第一个版本创建的视图还存在
 class DinDinSynchronous(models.TransientModel):
